Language.Lua.Syntax

Description

Abstract syntax of Lua 5.3 source files. See http://www.lua.org/manual/5.3/ for more information.

AST nodes

data Ident a

An identifier, defined as any string of letters, digits, or underscores, not beginning with a digit.

http://www.lua.org/manual/5.3/manual.html#3.1

Constructors

Ident !a !String

data IdentList1 a

One or more Idents.

Constructors

IdentList1 !a !(NonEmpty (Ident a))

type Chunk = Block

A chunk; Lua's compilation unit.

http://www.lua.org/manual/5.3/manual.html#3.3.2

data Block a

A block of statements, possibly ending in a return statement.

http://www.lua.org/manual/5.3/manual.html#3.3.1

Constructors

Block !a ![Statement a] !(Maybe (ReturnStatement a))

data Statement a

A statement.

http://www.lua.org/manual/5.3/manual.html#3.3

Constructors

EmptyStmt !a
;

Assign !a !(VariableList1 a) !(ExpressionList1 a)
var1, var2, var3 = exp1, exp2, exp3

FunCall !a !(FunctionCall a)
foo.bar(args)

Label !a !(Ident a)
::label::

Break !a
break

Goto !a !(Ident a)
goto label

Do !a !(Block a)
do block

While !a !(Expression a) !(Block a)
while exp do block end

Repeat !a !(Block a) !(Expression a)
repeat block until exp

If !a !(NonEmpty (Expression a, Block a)) !(Maybe (Block a))
if exp then block else block end

For !a !(Ident a) !(Expression a) !(Expression a) !(Maybe (Expression a)) !(Block a)
for x = exp do block end

ForIn !a !(IdentList1 a) !(ExpressionList1 a) !(Block a)
for a, b, c in exp1, exp2, exp3 do block end

FunAssign !a !(FunctionName a) !(FunctionBody a)
function name body

LocalFunAssign !a !(Ident a) !(FunctionBody a)
local function name body

LocalAssign !a !(IdentList1 a) !(ExpressionList a)
local x, y, z

data ReturnStatement a

Constructors

ReturnStatement !a !(ExpressionList a)
return exp1, exp2

data FunctionName a

Constructors

FunctionName !a !(IdentList1 a) !(Maybe (Ident a))
foo.bar:baz

data Variable a

A variable.

http://www.lua.org/manual/5.3/manual.html#3.2

Constructors

VarIdent !a !(Ident a)
x

VarField !a !(PrefixExpression a) !(Expression a)
table[exp]

VarFieldName !a !(PrefixExpression a) !(Ident a)
table.field

data VariableList1 a

One or more Variables.

Constructors

VariableList1 !a !(NonEmpty (Variable a))

data Expression a

An expression.

http://www.lua.org/manual/5.3/manual.html#3.4

Constructors

Nil !a

Bool !a !Bool

Integer !a !String

Float !a !String

String !a !String

Vararg !a

FunDef !a !(FunctionBody a)

PrefixExp !a !(PrefixExpression a)

TableCtor !a !(TableConstructor a)

Binop !a !(Binop a) !(Expression a) !(Expression a)

Unop !a !(Unop a) !(Expression a)

data ExpressionList a

Zero or more Expressions.

Constructors

ExpressionList !a ![Expression a]

data ExpressionList1 a

One or more Expressions.

Constructors

ExpressionList1 !a !(NonEmpty (Expression a))

data PrefixExpression a

Constructors

PrefixVar !a !(Variable a)

PrefixFunCall !a !(FunctionCall a)

Parens !a !(Expression a)

data FunctionCall a

A function call. May be a statement or an expression.

http://www.lua.org/manual/5.3/manual.html#3.3.6

http://www.lua.org/manual/5.3/manual.html#3.4.10

Constructors

FunctionCall !a !(PrefixExpression a) !(FunctionArgs a)

MethodCall !a !(PrefixExpression a) !(Ident a) !(FunctionArgs a)

data FunctionArgs a

Constructors

Args !a !(ExpressionList a)
(exp1, exp2)

ArgsTable !a !(TableConstructor a)
{ x = exp }

ArgsString !a !String
"str"

data FunctionBody a

Constructors

FunctionBody !a !(IdentList1 a) !Bool !(Block a)
(x, y, ...) block end

FunctionBodyVararg !a !(Block a)
(...) block end

data TableConstructor a

A table constructor.

http://www.lua.org/manual/5.3/manual.html#3.4.9

Constructors

TableConstructor !a !(FieldList a)
{ x = 5, [f(1)] = 6, 7 }

data Field a

Constructors

FieldExp !a !(Expression a) !(Expression a)
[exp1] = exp2

FieldIdent !a !(Ident a) !(Expression a)
name = exp

Field !a !(Expression a)
exp

data FieldList a

Zero or more Fields, separated by , or ;.

Constructors

FieldList !a ![Field a]

data Binop a

Constructors

Plus !a
+

Minus !a
-

Mult !a
*

FloatDiv !a
/

FloorDiv !a
//

Exponent !a
^

Modulo !a
%

BitwiseAnd !a
&

BitwiseXor !a
~

BitwiseOr !a
|

Rshift !a
>>

Lshift !a
<<

Concat !a
..

Lt !a
<

Leq !a
<=

Gt !a
>

Geq !a
>=

Eq !a
==

Neq !a
~=

And !a
and

Or !a
or

data Unop a

Constructors

Negate !a
-

Not !a
not

Length !a
#

BitwiseNot !a
~

Annotated typeclass

class Functor ast => Annotated ast where

Methods

ann :: Lens' (ast a) a